Try removing android specific conftest.
