+++ experimenter/tests/pytest.ini
@@ -2,6 +2,7 @@
 addopts = -p no:warnings
 log_cli = true
 log_cli_level = info
+reruns = 1
 markers =
     use_variables: marks tests that need to use pytest-variables
     run_targeting: run jexl targeting tests in firefox
